Move printing-related code to own directory
