+++ link-grammar/linkage.c
@@ -24,8 +24,8 @@
 #include "link-includes.h"
 #include "linkage.h"
 #include "post-process.h"
-#include "print.h"
-#include "print-util.h"
+#include "print/print.h"
+#include "print/print-util.h"
 #include "sat-solver/sat-encoder.h"
 #include "string-set.h"
 #include "structures.h"
