+++ link-grammar/tokenize.c
@@ -23,8 +23,8 @@
 #include "dict-common.h"
 #include "error.h"
 #include "externs.h"
-#include "print.h"
-#include "print-util.h"
+#include "print/print.h"
+#include "print/print-util.h"
 #include "regex-morph.h"
 #include "spellcheck.h"
 #include "string-set.h"
